removed single candidate controller
