+++ web/src/main/java/com/epam/rft/atsy/web/controllers/rest/CandidateController.java
@@ -8,28 +8,40 @@ import com.epam.rft.atsy.service.request.CandidateFilterRequest;
 import com.epam.rft.atsy.service.request.SearchOptions;
 import com.epam.rft.atsy.service.response.PagingResponse;
 import com.epam.rft.atsy.web.controllers.LogicallyDeletableAbstractController;
+import com.epam.rft.atsy.web.exceptionhandling.RestResponse;
+import com.epam.rft.atsy.web.helper.CandidateValidatorHelper;
 import com.epam.rft.atsy.web.messageresolution.MessageKeyResolver;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.FieldError;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.io.IOException;
+import java.util.Collections;
+import java.util.List;
+import java.util.Locale;
+import javax.validation.Valid;
 
 /**
  * REST controller used to retrieve information from the stored candidates in JSON format.
  * e.g.: It fills the table with data on the welcome page.
  */
 @RestController
-@RequestMapping(value = "/secure/candidates")
-public abstract class CandidateController
-    extends LogicallyDeletableAbstractController<CandidateDTO> {
+@RequestMapping(path = "/secure/candidate")
+public class CandidateController extends LogicallyDeletableAbstractController<CandidateDTO> {
 
+  private static final String COMMON_INVALID_INPUT_MESSAGE_KEY = "common.invalid.input";
+  private static final String CANDIDATE_ERROR_DUPLICATE_MESSAGE_KEY = "candidate.error.duplicate";
   private static final Logger LOGGER = LoggerFactory.getLogger(CandidateController.class);
 
   private static final SearchOptions
@@ -38,13 +50,15 @@ public abstract class CandidateController
           .phone(StringUtils.EMPTY).build();
 
   private CandidateService candidateService;
+  private CandidateValidatorHelper candidateValidatorHelper;
 
   @Autowired
   public CandidateController(
       CandidateService candidateService,
-      MessageKeyResolver messageKeyResolver) {
+      MessageKeyResolver messageKeyResolver, CandidateValidatorHelper candidateValidatorHelper) {
     super(candidateService, messageKeyResolver);
     this.candidateService = candidateService;
+    this.candidateValidatorHelper = candidateValidatorHelper;
   }
 
   /**
@@ -81,6 +95,67 @@ public abstract class CandidateController
             .candiadtePositions(position).build();
 
     return candidateService.getCandidatesByFilterRequest(candidateFilterRequest);
+  }
+
+  /**
+   * Saves or updates and existing candidate.
+   * @param candidateDTO an object which wraps the data of a candidate
+   * @param result an object used to check if any error occurs
+   * @param locale language of the response
+   * @return a ResponseEntity object, which contains HTTP status code and error message if it occurs
+   */
+  @Override
+  public ResponseEntity saveOrUpdate(@Valid @RequestBody CandidateDTO candidateDTO,
+                                     BindingResult result, Locale locale) {
+    if (!result.hasErrors()) {
+      Long candidateId = candidateService.saveOrUpdate(candidateDTO);
+      return new ResponseEntity<>(Collections.singletonMap("id", candidateId),
+          HttpStatus.OK);
+    } else {
+      RestResponse restResponse = parseValidationErrors(result.getFieldErrors(), locale);
+      return new ResponseEntity<>(restResponse, HttpStatus.BAD_REQUEST);
+    }
+  }
+
+  @RequestMapping(path = "/validate", method = RequestMethod.POST)
+  public ResponseEntity validateCandidate(@Valid @RequestBody CandidateDTO candidateDTO,
+                                          BindingResult result,
+                                          Locale locale) {
+    if (result.hasErrors()) {
+      return new ResponseEntity(parseValidationErrors(result.getFieldErrors(), locale),
+          HttpStatus.BAD_REQUEST);
+    }
+
+    // If candidate id is null, it means new candidate would like to be created
+    if (candidateDTO.getId() == null) {
+      return createResponseEntityForCandidateValidation(
+          this.candidateValidatorHelper.isValidNonExistingCandidate(candidateDTO));
+    } else {
+      return createResponseEntityForCandidateValidation(
+          this.candidateValidatorHelper.isValidExistingCandidate(candidateDTO));
+    }
+  }
+
+  private ResponseEntity createResponseEntityForCandidateValidation(boolean hasNoErrors) {
+    if (hasNoErrors) {
+      return new ResponseEntity(RestResponse.NO_ERROR, HttpStatus.OK);
+    }
+    String errorMessage =
+        this.messageKeyResolver.resolveMessageOrDefault(CANDIDATE_ERROR_DUPLICATE_MESSAGE_KEY);
+    return new ResponseEntity(new RestResponse(errorMessage), HttpStatus.BAD_REQUEST);
+  }
+
+  private RestResponse parseValidationErrors(List<FieldError> fieldErrors, Locale locale) {
+    String
+        errorMessage =
+        messageKeyResolver.resolveMessageOrDefault(COMMON_INVALID_INPUT_MESSAGE_KEY);
+    RestResponse restResponse = new RestResponse(errorMessage);
+
+    for (FieldError fieldError : fieldErrors) {
+      restResponse.addField(fieldError.getField(),
+          messageKeyResolver.resolveMessageOrDefault(fieldError.getDefaultMessage()));
+    }
+    return restResponse;
   }
 
   private SearchOptions parseFilters(String filterJSON) {
