+++ web/src/main/webapp/WEB-INF/pages/candidate_create.jsp
@@ -10,7 +10,7 @@
 <spring:url value="/secure/positions" var="positions"/>
 <spring:url value="/secure/channels" var="channels"/>
 <spring:url value="/secure/welcome" var="welcome"/>
-<spring:url value="/secure/candidate" var="candidateURL"/>
+<spring:url value="/secure/candidate/details" var="candidateURL"/>
 <spring:url value="/secure/application" var="application"/>
 
 <atsy:secure_page>
