+++ web/src/main/webapp/resources/js/atsy-candidate.js
@@ -75,7 +75,7 @@ $('.table').bootstrapTable({
       },
 
       onClickRow: function (row, $element) {
-        window.location.href = "candidate/" + row.id;
+        window.location.href = "candidate/details/" + row.id;
       }
 });
 
@@ -90,7 +90,7 @@ function actionsFormatter(value, row, index) {
 window.candidatesEvents = {
     'click .remove': function (e, value, row) {
          var container = $('#candidates_table');
-         var options = getOptions('question.delete.candidate.js', 'selected.candidate.not.found.js', row, container, "candidates");
+         var options = getOptions('question.delete.candidate.js', 'selected.candidate.not.found.js', row, container, "candidate");
          e.stopImmediatePropagation();
          bootbox.dialog(options);
     }
